aca-streamlit-admin/requirements.txt
streamlit==1.32.2
requests==2.31.0
azure-identity==1.16.1